fix: missing 18n and handling epics and stories delete

+++ pages/projects/[id]/tasks/-composables/useCreateTask.ts
@@ -11,14 +11,14 @@ type CreateTaskPayload = {
   title: string;
   organization: TaskSimple['organization'];
   parentTask?: number;
-  sprint?: number;
+  sprint?: SprintSimple;
 };
 
-type CreateTaskResponse = Omit<TaskSimple, 'parent_task'> & {
+type CreateTaskResponse = Omit<TaskSimple, 'parent_task' | 'sprint'> & {
   learningplan: number;
-  sprint: number;
   kanban_column: number;
   parent_task?: number;
+  sprint?: number;
 };
 
 export const useCreateTask = (
@@ -40,11 +40,16 @@ export const useCreateTask = (
         submission_required: false,
         title,
         parent_task: parentTask,
-        sprint,
+        sprint: sprint?.id || undefined,
         organization,
       });
-      task.data.sprint_id = sprint || 0;
-      return task.data as TaskSimple;
+
+      const taskWithSprintId = {
+        ...task.data,
+        sprint_id: sprint?.id || undefined,
+      };
+
+      return taskWithSprintId as TaskSimple & { sprint_id?: number };
     },
     onSuccess(data) {
       queryClient.setQueryData<SprintsResponse>(['sprints', learninplanId], (oldData) => {
@@ -72,25 +77,21 @@ export const useCreateTask = (
           };
         }
       });
-      setMessage(
-        t('pages.projects.tasks.actions.created_success', { item: t('pages.projects.common.task') }),
-        'success',
-        true,
-      );
+      setMessage(t('pages.projects.tasks.actions.created_success', { item: data.title }), 'success', true);
     },
-    onError() {
-      setMessage(
-        t('pages.projects.tasks.actions.created_error', { item: t('pages.projects.common.task') }),
-        'error',
-        true,
-      );
+    onError(_, variables) {
+      setMessage(t('pages.projects.tasks.actions.created_error', { item: variables.title }), 'error', true);
     },
   });
 
 type DeleteTaskPayload = {
+  title: string;
   id: number;
   sprintId?: number;
+  hasChildren?: boolean;
+  organization: TaskSimple['organization'];
 };
+
 export const useDeleteTask = (
   learninplanId: Ref<number>,
   queryClient: QueryClient,
@@ -98,7 +99,11 @@ export const useDeleteTask = (
   t: Function,
 ) =>
   useMutation({
-    mutationFn({ id }: DeleteTaskPayload) {
+    mutationFn({ id, organization, hasChildren }: DeleteTaskPayload) {
+      if (hasChildren) {
+        setMessage(t(`pages.projects.tasks.actions.delete_${organization}_has_children`), 'warning', true, false, true);
+        return Promise.reject(new Error('Item has children'));
+      }
       return strapi.delete('tasks', id);
     },
     onSuccess(_, variables) {
@@ -120,18 +125,13 @@ export const useDeleteTask = (
           backlog: oldData.backlog.filter((task) => task.id !== variables.id),
         };
       });
-      setMessage(
-        t('pages.projects.tasks.actions.deleted_success', { item: t('pages.projects.common.task') }),
-        'error',
-        true,
-      );
+      setMessage(t('pages.projects.tasks.actions.deleted_success', { item: variables.title }), 'success', true);
     },
-    onError() {
-      setMessage(
-        t('pages.projects.tasks.actions.deleted_error', { item: t('pages.projects.common.task') }),
-        'error',
-        true,
-      );
+    onError(error, variables) {
+      if (error.message === 'Item has children') {
+        throw error;
+      }
+      setMessage(t('pages.projects.tasks.actions.deleted_error', { item: variables.title }), 'error', true);
     },
   });
